+++ internal/schema/schema.go
@@ -403,73 +403,6 @@ func parseSchema(s *Schema, l *common.Lexer) error {
 	return nil
 }
 
-func validateEntryPointName(s *Schema, entryPoint *EntryPoint) error {
-	if s == Meta {
-		return nil
-	}
-	switch name := entryPoint.Name; name {
-	case "query", "mutation", "subscription":
-		if prev, ok := s.entryPointNames[name]; ok {
-			return &errors.QueryError{
-				Message:   fmt.Sprintf(`%q type provided more than once`, name),
-				Locations: []errors.Location{prev.Loc, entryPoint.Loc},
-			}
-		}
-	default:
-		return &errors.QueryError{
-			Message:   fmt.Sprintf(`unexpected %q, expected "query", "mutation" or "subscription"`, name),
-			Locations: []errors.Location{entryPoint.Loc},
-		}
-	}
-	return nil
-}
-
-func validateTypeName(s *Schema, t NamedType) error {
-	if s == Meta {
-		return nil
-	}
-	name := t.TypeName()
-	if strings.HasPrefix(name, "__") {
-		return &errors.QueryError{
-			Message:   fmt.Sprintf(`%q must not begin with "__", reserved for introspection types`, name),
-			Locations: []errors.Location{t.Location()},
-		}
-	}
-	if _, ok := Meta.Types[name]; ok {
-		return &errors.QueryError{
-			Message:   fmt.Sprintf(`built-in type %q redefined`, name),
-			Locations: []errors.Location{t.Location()},
-		}
-	}
-	if prev, ok := s.Types[name]; ok {
-		return &errors.QueryError{
-			Message:   fmt.Sprintf(`%q defined more than once`, name),
-			Locations: []errors.Location{prev.Location(), t.Location()},
-		}
-	}
-	return nil
-}
-
-func validateDirectiveName(s *Schema, directive *DirectiveDecl) error {
-	if s == Meta {
-		return nil
-	}
-	name := directive.Name
-	if _, ok := Meta.Directives[name]; ok {
-		return &errors.QueryError{
-			Message:   fmt.Sprintf(`built-in directive %q redefined`, name),
-			Locations: []errors.Location{directive.Loc},
-		}
-	}
-	if prev, ok := s.Directives[name]; ok {
-		return &errors.QueryError{
-			Message:   fmt.Sprintf("%q defined more than once", name),
-			Locations: []errors.Location{prev.Loc, directive.Loc},
-		}
-	}
-	return nil
-}
-
 func parseObjectDecl(l *common.Lexer) *Object {
 	o := &Object{}
 	ident := l.ConsumeIdentWithLoc()
